Separate Ingest Service
This patch splits up the ingest and capture parts of pyCA. This will
allow simultaneous ingests and recordings. It should also make it easy
to add re-ingest functionality for failed ingests.
