+++ pyca/utils.py
@@ -9,6 +9,7 @@
 '''
 
 from pyca.config import config
+from pyca import db
 from datetime import datetime
 from dateutil.tz import tzutc
 import errno
@@ -137,3 +138,64 @@ def configure_service(service):
 def ensurelist(x):
     '''Ensure an element is a list.'''
     return x if type(x) == list else [x]
+
+
+def register_ca(status='idle'):
+    '''Register this capture agent at the Matterhorn admin server so that it
+    shows up in the admin interface.
+
+    :param address: Address of the capture agent web ui
+    :param status: Current status of the capture agent
+    '''
+    # If this is a backup CA we don't tell the Matterhorn core that we are
+    # here.  We will just run silently in the background:
+    if config()['agent']['backup_mode']:
+        return True
+    params = [('address', config()['ui']['url']), ('state', status)]
+    url = '%s/agents/%s' % (config()['service-capture.admin'][0],
+                            config()['agent']['name'])
+    try:
+        response = http_request(url, params).decode('utf-8')
+        if response:
+            logging.info(response)
+    except:
+        # Ignore errors (e.g. network issues) as it's more important to get
+        # the recording as to set the correct current state in the admin ui.
+        logging.warning('Could not set capture agent state')
+        logging.warning(traceback.format_exc())
+        return False
+    return True
+
+
+def recording_state(recording_id, status):
+    '''Send the state of the current recording to the Matterhorn core.
+
+    :param recording_id: ID of the current recording
+    :param status: Status of the recording
+    '''
+    # If this is a backup CA we do not update the recording state since the
+    # actual CA does that and we want to interfere.  We will just run silently
+    # in the background:
+    if config()['agent']['backup_mode']:
+        return
+    params = [('state', status)]
+    url = config()['service-capture.admin'][0]
+    url += '/recordings/%s' % recording_id
+    try:
+        result = http_request(url, params)
+        logging.info(result)
+    except:
+        # Ignore errors (e.g. network issues) as it's more important to get
+        # the recording as to set the correct current state in the admin ui.
+        logging.warning('Could not set recording state')
+        logging.warning(traceback.format_exc())
+
+
+def update_event_status(event, status):
+    '''Update the status of a particular event in the database.
+    '''
+    dbs = db.get_session()
+    dbs.query(db.RecordedEvent).filter(db.RecordedEvent.start == event.start)\
+                               .update({'status': status})
+    event.status = status
+    dbs.commit()
